File descriptions to allow multiple area files
In case of multiple area files with the same load, "_" (underscore) can be used to describe multiple files.
Description needs to be after the load, which is a bit unfortunate for viewing in file explorer. Maybe subject to change in the future.
Simplified splitting of the path compared to the original.